Create 1- Remove Duplicates
This problem can be solve with recurion. Here we are finding ans for starting index and leaving the rest on recursion. This is also called recursion leap of faith.  Here we have to work on consecutive elements and have to do this repeatdly ,so recursion can be a effective way to solve this problem.